mv cli_setup to airtest.cli.parser
