+++ playground/blackjack_example/test_blackjack.py
@@ -3,47 +3,14 @@
 __author__ = "刘欣"
 
 from airtest.core.api import *
-from airtest.report.report import simple_report
+from airtest.cli.parser import cli_setup
 
 
 PKG = "org.cocos2d.blackjack"
 APK = "blackjack-release-signed.apk"
 
-def cli_setup(args=None):
-    from airtest.cli.parser import runner_parser
-    from airtest.cli.runner import setup_by_args
-    import argparse
-    import sys
-
-    if not args:
-        if len(sys.argv) < 2:
-            print("no cmdline args")
-            return False
-        args = sys.argv
-    print(args)
-
-    ap = argparse.ArgumentParser()
-    if "--report" in args:
-        from airtest.report.report import main as report_main
-        from airtest.report.report import get_parger as report_parser
-        ap = report_parser(ap)
-        args = ap.parse_args(args)
-        report_main(args)
-        exit(0)
-    else:
-        ap = runner_parser(ap)
-        args = ap.parse_args(args)
-        setup_by_args(args)
-    return True
-
 
 if not cli_setup():
-    # set log file if you want a html report
-    # set_logdir("log")
-    # connect android device with params: cap_method=javacap
-    # init_device("Android", cap_method="javacap")
-
-    # or setup env at once
     auto_setup(
         devices=["Android:///?cap_method=javacap&ori_method=adbori"],
         logdir="log",
@@ -77,4 +44,5 @@ snapshot(msg="app stopped")
 print("test finished")
 
 # generate html report
+from airtest.report.report import simple_report
 simple_report("log", __file__)
